+++ app/helpers/schedules_helper.rb
@@ -1,0 +1,14 @@
+module SchedulesHelper
+  def schedule_category_color(schedule)
+    case schedule.category
+    when 'coaching'
+      'bg-paleFern'
+    when 'qa'
+      'bg-paleCoral'
+    when 'meeting'
+      'bg-gold'
+    else
+      'bg-coral'
+    end
+  end
+end