+++ app/javascript/controllers/popover_controller.js
@@ -1,0 +1,8 @@
+import Popover from 'stimulus-popover'
+
+export default class extends Popover {
+  connect() {
+    super.connect()
+    console.log('Do what you want here.')
+  }
+}