+++ db/migrate/20230919104523_create_schedules.rb
@@ -1,0 +1,13 @@
+class CreateSchedules < ActiveRecord::Migration[7.0]
+  def change
+    create_table :schedules do |t|
+      t.string :title
+      t.integer :category
+      t.datetime :start_time
+      t.datetime :end_time
+      t.references :user, null: false, foreign_key: true
+
+      t.timestamps
+    end
+  end
+end