+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_09_19_013932) do
+ActiveRecord::Schema[7.0].define(version: 2023_09_21_030626) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -184,6 +184,25 @@ ActiveRecord::Schema[7.0].define(version: 2023_09_19_013932) do
     t.index ["user_id"], name: "index_qualities_on_user_id"
   end
 
+  create_table "schedules", force: :cascade do |t|
+    t.string "title"
+    t.integer "category"
+    t.datetime "start_time"
+    t.datetime "end_time"
+    t.bigint "user_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "link"
+    t.index ["user_id"], name: "index_schedules_on_user_id"
+  end
+
+  create_table "schedules_users", id: false, force: :cascade do |t|
+    t.bigint "schedule_id", null: false
+    t.bigint "user_id", null: false
+    t.index ["schedule_id", "user_id"], name: "index_schedules_users_on_schedule_id_and_user_id"
+    t.index ["user_id", "schedule_id"], name: "index_schedules_users_on_user_id_and_schedule_id"
+  end
+
   create_table "ticket_details", force: :cascade do |t|
     t.text "content"
     t.integer "access_level"
@@ -250,6 +269,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_09_19_013932) do
   add_foreign_key "qa_templates", "accounts"
   add_foreign_key "qualities", "accounts"
   add_foreign_key "qualities", "users"
+  add_foreign_key "schedules", "users"
   add_foreign_key "ticket_details", "tickets"
   add_foreign_key "tickets", "accounts"
   add_foreign_key "tickets", "users"