+++ Changelog.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 
+## [0.13.0] : 2023-21-09
+
+### Added
+
+- Added calendar and schedule
+- Added schedule Notification and Mailer
+- Migrated joinable table schedules_users
+
 ## [0.12.0] : 2023-19-09
 
 ### Added
@@ -126,6 +134,7 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 [unreleased]: https://github.com/danrayfr/unified/branch/development#diff
+[0.13.0]: https://github.com/danrayfr/unified/pull/21
 [0.12.0]: https://github.com/danrayfr/unified/pull/20
 [0.11.0]: https://github.com/danrayfr/unified/pull/19
 [0.10.2]: https://github.com/danrayfr/unified/pull/18